+++ src/index.ts
@@ -5,5 +5,6 @@ import BigipConfig from './ltm';
 export * from './ltm';
 export * from './models';
 export { RegExTree } from './regex';
+export { parseConfig, parseConfigs } from './universalParse';
 
 export default BigipConfig;